% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grant}
\alias{grant}
\title{A dataset that contains the information about grants opportunity details}
\format{
A data frame with 74669 rows and 4 variables
\describe{
\item{opportunity_id}{Integer ID for this opportunity}
\item{opportunity_title}{Title of the opportunity}
\item{expected_number_of_awards}{Expected count of awards}
\item{estimated_funding}{Estimated funding amount in dollars}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2023/2023-10-03}
}
\usage{
grant
}
\description{
A dataset that contains the information about grants opportunity details
}
\keyword{datasets}
