% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_grant_app}
\alias{run_grant_app}
\title{The grant explore app}
\usage{
run_grant_app()
}
\description{
The app is aimed to provide users an initial exploration about the possible grants they could apply by
filtering different categories, years and eligibility. They could also set the award dollar range to
meet their funding expectations.
}
\examples{
\dontrun{
run_grant_app()
}
}
